Project 4: Machine Language
To get a taste of low-level programming in machine language, and to get
acquainted with the Hack computer platform. In the process of working on
this project, you will become familiar with the assembly process -
translating from symbolic language to machine-language - and you will
appreciate visually how native binary code executes on the target
hardware platform. These lessons will be learned in the context of
writing and testing two low-level programs, as follows.